vo: tidy code to reduce warnings
Mostly get rid of raw types, also remove some redundant typecasts.

This includes slapping a SuppressWarnings annontations on a couple
of classes that use Swing classes that have acquired a parameterised
type between java 6 and java 8, so that the classes can still build
under either version but don't issue warnings at java 8.

Also fix various javadoc errors.

+++ vo/src/main/uk/ac/starlink/vo/UwsJob.java
@@ -794,7 +794,7 @@ public class UwsJob {
         byte[] bbuf = new byte[ nc ];
         for ( int i = 0; i < nc; i++ ) {
             char c = sbuf.charAt( i );
-            assert (int) c == (int) ( c & 0x7f );
+            assert (int) c == ( c & 0x7f );
             bbuf[ i ] = (byte) sbuf.charAt( i );
         }
         return bbuf;
